Bring back buildifer_availability
