+++ src/buildifier/buildifier.ts
@@ -22,7 +22,6 @@ import {
   getBuildifierExecutablePath,
   getBuildifierJsonConfigPath,
 } from "../extension/configuration";
-import { logDebug, logWarn } from "../extension/logger";
 
 const execFile = util.promisify(child_process.execFile);
 type PromiseExecFileException = child_process.ExecFileException & {
@@ -120,34 +119,6 @@ export async function buildifierLint(
   }
 }
 
-/**
- * Validates that buildifier executable is working correctly.
- *
- * @param buildifierExecutable Custom executable path to use instead of default.
- * @returns True if buildifier is working correctly, false otherwise.
- */
-export async function validateBuildifierExecutable(
-  buildifierExecutable: string,
-): Promise<boolean> {
-  try {
-    logDebug(`Testing buildifier with JSON output format`);
-    const { stdout } = await executeBuildifier(
-      "",
-      ["--format=json", "--mode=check", "--lint=off"],
-      false,
-      buildifierExecutable,
-    );
-    JSON.parse(stdout); // Will throw if not valid JSON
-    logDebug(`Buildifier validation successful - JSON output supported`);
-    return true;
-  } catch (jsonError) {
-    logWarn(
-      `Buildifier JSON validation failed. The buildifier version may be too old and doesn't support JSON output format. Consider updating to a newer version. Error: ${jsonError instanceof Error ? jsonError.message : String(jsonError)}`,
-      false,
-    );
-    return false;
-  }
-}
 
 /**
  * Executes buildifier with the given file content and arguments.
