+++ src/buildifier/index.ts
@@ -13,6 +13,7 @@
 // limitations under the License.
 
 export * from "./buildifier";
+export * from "./buildifier_availability";
 export * from "./buildifier_diagnostics_manager";
 export * from "./buildifier_format_provider";
 export * from "./buildifier_result";
